requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
mysql-connector-python==9.3.0
